wip: Refactor native code.
