+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/string_helpers.h
@@ -1,3 +1,25 @@
+/*
+ * Copyright (c) 2021 PlayEveryWare
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 #pragma once
 
 namespace playeveryware::eos::string_helpers
@@ -53,6 +75,30 @@ namespace playeveryware::eos::string_helpers
         return result;
     }
 
+    //-------------------------------------------------------------------------
+    inline static bool create_timestamp_str(char* final_timestamp, size_t final_timestamp_len)
+    {
+        constexpr size_t buffer_len = 32;
+        char buffer[buffer_len];
+
+        if (buffer_len > final_timestamp_len)
+        {
+            return false;
+        }
+
+        time_t raw_time = time(NULL);
+        tm time_info = { 0 };
+
+        timespec time_spec = { 0 };
+        timespec_get(&time_spec, TIME_UTC);
+        localtime_s(&time_info, &raw_time);
+
+        strftime(buffer, buffer_len, "%Y-%m-%dT%H:%M:%S", &time_info);
+        long milliseconds = (long)round(time_spec.tv_nsec / 1.0e6);
+        snprintf(final_timestamp, final_timestamp_len, "%s.%03ld", buffer, milliseconds);
+
+        return true;
+    }
 
     //-------------------------------------------------------------------------
     inline static size_t utf8_str_bytes_required_for_wide_str(const wchar_t* wide_str, int wide_str_len = -1)
